requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-dotenv==1.0.0
google-generativeai==0.3.0
youtube-transcript-api==0.6.0
jinja2==3.1.2
google-auth==2.20.0
google-auth-oauthlib==0.5.3
